correct class String ,create and fill class Tickets, add some fixes to the other classes

+++ projectticket/projectticket/Row.cpp
@@ -8,7 +8,7 @@ Row::Row()
 Row::Row(int rowNumber, int numberOfSeats)
 {
     this->rowNumber = rowNumber;
-    for (size_t i = 0; i < numberOfSeats; i++)
+    for (int i = 0; i < numberOfSeats; i++)
     {
         Seat seat(rowNumber, i);
         seats.pushBack(seat);
